+++ maths/basics/ft_max.c
@@ -12,14 +12,14 @@
 
 #include "libft.h"
 
-size_t	ft_maxui(size_t a, size_t b)
+uint64_t	ft_maxui(uint64_t a, uint64_t b)
 {
 	if (a >= b)
 		return (a);
 	return (b);
 }
 
-uint64_t	ft_maxi(uint64_t a, uint64_t b)
+int64_t	ft_maxi(int64_t a, int64_t b)
 {
 	if (a >= b)
 		return (a);
