+++ maths/basics/ft_min.c
@@ -12,14 +12,14 @@
 
 #include "libft.h"
 
-size_t	ft_minui(size_t a, size_t b)
+uint64_t	ft_minui(uint64_t a, uint64_t b)
 {
 	if (a <= b)
 		return (a);
 	return (b);
 }
 
-uint64_t	ft_mini(uint64_t a, uint64_t b)
+int64_t	ft_mini(int64_t a, int64_t b)
 {
 	if (a <= b)
 		return (a);
